pyproject.toml
[project]
name = "europa"
version = "0.0.20"
description = "Europa - AI coordinator powered by Google Gemini with MCP integration"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent"
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "mcp==1.9.3",
    "opentelemetry-distro>=0.50b0",
    "opentelemetry-exporter-otlp-proto-http>=1.29.0",
    "pydantic-settings>=2.7.0",
    "pydantic[email]>=2.10.4",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "typer>=0.15.1",
    "anthropic>=0.49.0",
    "openai>=1.63.2",
    "prompt-toolkit>=3.0.50",
    "aiohttp>=3.11.13",
    "a2a-types>=0.1.0",
    "opentelemetry-instrumentation-openai>=0.0.40.7; python_version >= '3.10' and python_version < '4.0'",
    "opentelemetry-instrumentation-anthropic>=0.40.7; python_version >= '3.10' and python_version < '4.0'",
    "opentelemetry-instrumentation-mcp>=0.40.7; python_version >= '3.10' and python_version < '4.0'",
    "google-genai",
    "opentelemetry-instrumentation-google-genai>=0.2b0",
    "tensorzero>=2025.4.7",
    "opentelemetry-instrumentation-google-genai>=0.2b0",
    "deprecated>=1.2.18",
    "pymongo>=4.13.1",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "google-auth",
    "uvicorn>=0.34.3",
    "requests>=2.32.3",
    "uvloop>=0.21.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.58.1",
]

dev = [
    "anthropic>=0.42.0",
    "pre-commit>=4.0.1",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "ruff>=0.8.4",
    "tomli>=2.2.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_agent"]

[[tool.poetry.packages]]
include = "mcp_agent"
from = "src"

[tool.hatch.build]
include = [
    "src/mcp_agent/**/*.py",
    "src/mcp_agent/**/*.yaml",
    "src/mcp_agent/**/*.yml",
    "src/mcp_agent/**/*.csv",
    "src/mcp_agent/**/*.txt",    
    "src/mcp_agent/**/*.example",    
    "examples/**/*.py",
    "examples/**/*.yaml",
    "examples/**/*.yml",
    "examples/**/*.txt",
    "examples/**/*.csv",
    "examples/**/*.md",
    "examples/**/*.css",
    "examples/**/mount-point/*.csv",
    "examples/workflows/short_story.txt",
    "examples/mcp/**/*.py",
    "examples/mcp/**/*.yaml",
    "examples/mcp/**/*.example",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "e2e: tests that connect to external resources (llms)",
    "integration: marks tests as integration tests",
    "simulated_endpoints: marks tests that use simulated external endpoints"
]

[dependency-groups]
dev = [
    "anthropic>=0.49.0",
    "pre-commit>=4.0.1",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "ruff>=0.8.4",
    "tomli>=2.2.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=6.1.1",
    "ipdb>=0.13.13",
]

[project.scripts]
europa = "mcp_agent.cli.__main__:app"
prompt-server = "mcp_agent.mcp.prompts.__main__:main"

[tool.setuptools.package-data]
mcp_agent = []

[tool.ruff]
line-length = 140
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "TCH"  # type checking
]
ignore = ["E501"]

# More lenient settings for existing code
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]  # Don't require type annotations in tests
